requirements.txt
grpcio==1.72.1
grpcio-tools==1.72.1
protobuf==6.31.1
setuptools==80.9.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
python-dotenv==1.0.0
